Adding a First Person Player
The controls for the player are made with Unity's Input System.
I've created a new scene to test the first person movement of the player and the camera.
I'm using a Cinemachine's camera that is following an empty object positioned in the players "head".
Using Cinemachine input Provider (Script) to make the camera follow the XY Axis - using the LookMovement input - using the mouse.